requirements.txt
streamlit
requests
schedule
python-dotenv
google-generativeai
